Split sql
